requirements.txt
python-dotenv==1.0.0
oracledb==1.2.2